Added services screen. Added staff creation.

+++ src/views/Staff.vue
@@ -7,6 +7,41 @@
       >
         <div class="row">
           <div class="col-2">
+            <h3>Radnici</h3>
+            <button @click="isCreateDropdownOpen = !isCreateDropdownOpen">
+              +
+            </button>
+            <div v-if="isCreateDropdownOpen">
+              <label for="create-name">
+                Ime
+              </label>
+              <input
+                id="create-name"
+                v-model="newStaff.name"
+                type="text"
+                name="name"
+              >
+              <label for="create-name">
+                Email
+              </label>
+              <input
+                id="create-email"
+                v-model="newStaff.email"
+                type="email"
+                name="email"
+              >
+              <button
+                :class="{
+                  btn: true,
+                  'btn-primary': !requestSent,
+                  'btn-success': requestSent && status,
+                  'btn-danger': requestSent && !status,
+                }"
+                @click="createStaff()"
+              >
+                Spremi
+              </button>
+            </div>
             <div class="tab">
               <ul
                 class="nav nav-tabs flex-column"
@@ -59,12 +94,15 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, ref } from 'vue';
+import {
+  computed, defineComponent, reactive, ref,
+} from 'vue';
 import { useStore } from '@/store';
 import Dashboard from '@/components/layout/Dashboard.vue';
 import StaffEdit from '@/components/staff/StaffEdit.vue';
 import ActionTypes from '@/store/action-types';
 import Staff from '@/types/staff';
+import { WorkingHours } from '@/types/workingHours';
 
 export default defineComponent({
   components: {
@@ -75,15 +113,54 @@ export default defineComponent({
     const store = useStore();
     const staff = computed(() => store.state.staff.allStaff);
     const selectedWorker = ref(staff.value[0].id);
+    const selectedCompany = computed(() => store.state.shared.selectedCompany);
+    const isCreateDropdownOpen = ref(false);
+    const requestSent = ref(false);
+    const status = ref(false);
+    const newStaff = reactive({
+      name: '',
+      email: '',
+      company: selectedCompany.value?.id,
+      hours: {
+        monday: { active: true, shifts: [{ start: '08:00', end: '16:00' }] },
+        tuesday: { active: true, shifts: [{ start: '08:00', end: '16:00' }] },
+        wednesday: { active: true, shifts: [{ start: '08:00', end: '16:00' }] },
+        thursday: { active: true, shifts: [{ start: '08:00', end: '16:00' }] },
+        friday: { active: true, shifts: [{ start: '08:00', end: '16:00' }] },
+        saturday: { active: false, shifts: [] },
+        sunday: { active: false, shifts: [] },
+      } as WorkingHours,
+    });
 
     function deleteWorker(worker: Staff) {
-      store.dispatch(ActionTypes.DELETE_STAFF, worker);
+      // eslint-disable-next-line no-alert
+      if (window.confirm('Jeste li sigurni da želite obrisati ovog radnika? Svi povezani termini također će biti izbrisani.')) {
+        store.dispatch(ActionTypes.DELETE_STAFF, worker);
+      }
+    }
+
+    async function createStaff() {
+      try {
+        await store.dispatch(ActionTypes.CREATE_STAFF, newStaff as Staff);
+        requestSent.value = true;
+        status.value = true;
+        newStaff.name = '';
+        newStaff.email = '';
+      } catch {
+        requestSent.value = true;
+        status.value = false;
+      }
     }
 
     return {
       staff,
       selectedWorker,
       deleteWorker,
+      isCreateDropdownOpen,
+      newStaff,
+      requestSent,
+      status,
+      createStaff,
     };
   },
 });
